refactor(api): split activity generation workflow by lesson kind
Turn activityGenerationWorkflow into a thin router that delegates to
core, custom, or language sub-workflows based on lesson kind. Fix custom
workflow saving before image generation completed.
