+++ apps/api/src/workflows/activity-generation/_utils/settled.ts
@@ -1,0 +1,3 @@
+export function settled<T>(result: PromiseSettledResult<T>, fallback: T): T {
+  return result.status === "fulfilled" ? result.value : fallback;
+}
