+++ apps/api/src/workflows/activity-generation/language-activity-workflow.test.ts
@@ -1,0 +1,146 @@
+import { randomUUID } from "node:crypto";
+import { generateActivityBackground } from "@zoonk/ai/tasks/activities/core/background";
+import { generateActivityCustom } from "@zoonk/ai/tasks/activities/custom";
+import { prisma } from "@zoonk/db";
+import { activityFixture } from "@zoonk/testing/fixtures/activities";
+import { chapterFixture } from "@zoonk/testing/fixtures/chapters";
+import { courseFixture } from "@zoonk/testing/fixtures/courses";
+import { lessonFixture } from "@zoonk/testing/fixtures/lessons";
+import { aiOrganizationFixture } from "@zoonk/testing/fixtures/orgs";
+import { beforeAll, beforeEach, describe, expect, test, vi } from "vitest";
+import { activityGenerationWorkflow } from "./activity-generation-workflow";
+
+vi.mock("workflow", () => ({
+  FatalError: class FatalError extends Error {},
+  getWorkflowMetadata: vi.fn().mockReturnValue({ workflowRunId: "test-run-id" }),
+  getWritable: vi.fn().mockReturnValue({
+    getWriter: () => ({
+      releaseLock: vi.fn(),
+      write: vi.fn().mockResolvedValue(null),
+    }),
+  }),
+  workflowStep: vi.fn().mockImplementation((_name: string, fn: unknown) => fn),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/background", () => ({
+  generateActivityBackground: vi.fn().mockResolvedValue({
+    data: { steps: [{ text: "Background step 1 text", title: "Background Step 1" }] },
+  }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/explanation", () => ({
+  generateActivityExplanation: vi.fn().mockResolvedValue({ data: { steps: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/mechanics", () => ({
+  generateActivityMechanics: vi.fn().mockResolvedValue({ data: { steps: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/examples", () => ({
+  generateActivityExamples: vi.fn().mockResolvedValue({ data: { steps: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/story", () => ({
+  generateActivityStory: vi.fn().mockResolvedValue({ data: { steps: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/challenge", () => ({
+  generateActivityChallenge: vi.fn().mockResolvedValue({ data: { steps: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/review", () => ({
+  generateActivityReview: vi.fn().mockResolvedValue({ data: { questions: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/core/explanation-quiz", () => ({
+  generateActivityExplanationQuiz: vi.fn().mockResolvedValue({ data: { questions: [] } }),
+}));
+
+vi.mock("@zoonk/ai/tasks/activities/custom", () => ({
+  generateActivityCustom: vi.fn().mockResolvedValue({
+    data: { steps: [{ text: "Custom step 1 text", title: "Custom Step 1" }] },
+  }),
+}));
+
+vi.mock("@zoonk/ai/tasks/steps/visual", () => ({
+  generateStepVisuals: vi.fn().mockResolvedValue({ data: { visuals: [] } }),
+}));
+
+vi.mock("@zoonk/core/steps/visual-image", () => ({
+  generateVisualStepImage: vi.fn().mockResolvedValue({ data: null, error: null }),
+}));
+
+vi.mock("@zoonk/core/steps/image", () => ({
+  generateStepImage: vi.fn().mockResolvedValue({ data: null, error: null }),
+}));
+
+vi.mock("@zoonk/core/cache/revalidate", () => ({
+  revalidateMainApp: vi.fn().mockResolvedValue(null),
+}));
+
+describe("language activity workflow", () => {
+  let organizationId: number;
+  let course: Awaited<ReturnType<typeof courseFixture>>;
+  let chapter: Awaited<ReturnType<typeof chapterFixture>>;
+
+  beforeAll(async () => {
+    const org = await aiOrganizationFixture();
+    organizationId = org.id;
+    course = await courseFixture({ organizationId });
+    chapter = await chapterFixture({
+      courseId: course.id,
+      organizationId,
+      title: `Test Chapter ${randomUUID()}`,
+    });
+  });
+
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  test("routes to language workflow (no-op) without calling any AI tasks", async () => {
+    const testLesson = await lessonFixture({
+      chapterId: chapter.id,
+      kind: "language",
+      organizationId,
+      title: `Language Lesson ${randomUUID()}`,
+    });
+
+    await activityFixture({
+      generationStatus: "pending",
+      kind: "background",
+      lessonId: testLesson.id,
+      organizationId,
+      title: `Background ${randomUUID()}`,
+    });
+
+    await activityGenerationWorkflow(testLesson.id);
+
+    expect(generateActivityBackground).not.toHaveBeenCalled();
+    expect(generateActivityCustom).not.toHaveBeenCalled();
+  });
+
+  test("activities remain pending after language workflow (no-op)", async () => {
+    const testLesson = await lessonFixture({
+      chapterId: chapter.id,
+      kind: "language",
+      organizationId,
+      title: `Language Pending Lesson ${randomUUID()}`,
+    });
+
+    const activity = await activityFixture({
+      generationStatus: "pending",
+      kind: "background",
+      lessonId: testLesson.id,
+      organizationId,
+      title: `Background ${randomUUID()}`,
+    });
+
+    await activityGenerationWorkflow(testLesson.id);
+
+    const dbActivity = await prisma.activity.findUnique({
+      where: { id: activity.id },
+    });
+    expect(dbActivity?.generationStatus).toBe("pending");
+  });
+});
